Modified id type to Long.

+++ src/main/java/com/pika/Onlinegrocerystore/Services/CustomerService.java
@@ -5,11 +5,11 @@ import com.pika.Onlinegrocerystore.domain.Customer;
 import java.util.List;
 
 public interface CustomerService {
-    Customer findById(int id);
+    Customer findById(Long id);
     List<Customer> findAll();
     void update(Customer customer);
     void create(Customer customer);
-    void deleteById(int id);
+    void deleteById(Long id);
     List<Customer> findByName(String name);
     Customer checkLogin(String name, String password);
 }
